+++ CHANGELOG.md
@@ -25,6 +25,8 @@ The format is based on [Keep a Changelog].
 * Indent level support when using yq to format JSON, XML, and YAML ([#259]).
 
 ### Formatters
+* `dhall-format` official formatter for
+  [Dhall](https://dhall-lang.org/) ([#284]).
 * [`dprint`](https://dprint.dev) for various (depending on
   [installed plugins](https://dprint.dev/plugins/)) ([#209]).
 * [`js-beautify`](https://github.com/beautify-web/js-beautify) for
@@ -69,6 +71,7 @@ The format is based on [Keep a Changelog].
 [#275]: https://github.com/radian-software/apheleia/pull/275
 [#279]: https://github.com/radian-software/apheleia/pull/279
 [#282]: https://github.com/radian-software/apheleia/pull/282
+[#284]: https://github.com/radian-software/apheleia/pull/284
 
 ## 4.0 (released 2023-11-23)
 ### Breaking changes